Safer queries/query errors for the db.
