+++ db/dbMod.ts
@@ -1,2 +1,3 @@
 export * from "./client.ts";
 export * from "./dbQueries.ts";
+export * from "./safeQueries.ts";
