+++ models/userModel.ts
@@ -4,13 +4,16 @@ import {
   queryAll,
   queryOne,
   safeQuery,
+  safeQueryAll,
+  safeExecute,
+  safeExecuteTransaction,
 } from "../db/dbMod.ts";
 
 export function createUser(username: string, email: string) {
   const query = `
       INSERT INTO users (username, email) VALUES (?, ?);
     `;
-  return execute(query, username, email);
+  return safeExecute(() => execute(query, username, email));
 }
 
 // Neccessary safe query function
@@ -18,7 +21,7 @@ export function getAllUsers() {
   const query = `
       SELECT * FROM users;
     `;
-  return queryAll(query);
+  return safeQueryAll(() => queryAll(query));
 }
 
 // export function getUserByUsername(username: string) {
@@ -45,7 +48,7 @@ export function updateUser(userId: number, username: string, email: string) {
   const query = `
   UPDATE users SET username = ?, email = ? WHERE id = ?;
   `;
-  return execute(query, username, email, userId);
+  return safeExecute(() => execute(query, username, email, userId));
 }
 
 // Neccessary safe query function
@@ -53,7 +56,7 @@ export function deleteUser(userId: number) {
   const query = `
       DELETE FROM users WHERE id = ?;
       `;
-  return execute(query, userId);
+  return safeExecute(() => execute(query, userId));
 }
 
 export function getUserAndUpdateUser(
@@ -66,5 +69,5 @@ export function getUserAndUpdateUser(
     "UPDATE users SET username = ?, email = ? WHERE id = ?",
   ];
   const params = [[userId], [username, email, userId]];
-  return executeTransaction(queries, params);
+  return safeExecuteTransaction(() => executeTransaction(queries, params));
 }
